requirements.txt
boto3==1.26.51
botocore==1.29.51
PyYAML==6.0
pyspark==3.4.0